backend/requirements.txt
# requirements.txt

fastapi
uvicorn[standard]
python-dotenv
langchain-groq
langchain-huggingface
pypdf2
gtts
python-multipart